+++ test/mocks/com_mock.cpp
@@ -65,7 +65,7 @@ void Com::send_reqLine(const uint8_t lineNumber, Err_t error) const {
   gComMock->send_reqLine(lineNumber, error);
 }
 
-void Com::send_indState(Carriage_t carriage, uint8_t position,
+void Com::send_indState(Carriage_t carriage, int16_t position,
                         Err_t error) const {
   assert(gComMock != nullptr);
   gComMock->send_indState(carriage, position, error);
