+++ test/mocks/encoders_mock.h
@@ -35,7 +35,7 @@ public:
   MOCK_METHOD0(getCarriage, Carriage_t());
   MOCK_METHOD0(getMachineType, Machine_t());
   MOCK_METHOD0(encA_interrupt, void());
-  MOCK_METHOD0(getPosition, uint8_t());
+  MOCK_METHOD0(getPosition, int16_t());
   MOCK_METHOD0(getHallActive, Direction_t());
   MOCK_METHOD1(getHallValue, uint16_t(Direction_t));
 };
